[warning package rename] - add ability to stream scan/query
