+++ fs2-aws-dynamodb/src/main/scala/fs2/aws/dynamodb/stream/parsers/BeforeAfter.scala
@@ -1,4 +1,4 @@
-package fs2.aws.dynamodb.parsers
+package fs2.aws.dynamodb.stream.parsers
 
 sealed trait DynamoDBBeforeAfter[T]
