+++ fs2-aws-dynamodb/src/main/scala/fs2/aws/dynamodb/stream/parsers/package.scala
@@ -1,4 +1,4 @@
-package fs2.aws.dynamodb
+package fs2.aws.dynamodb.stream
 
 import java.util
